requirements.txt
anthropic>=0.39.0
feedparser>=6.0.11
PyYAML>=6.0.1
python-dotenv>=1.0.1
Jinja2>=3.1.4
beautifulsoup4>=4.12.3
requests>=2.32.3
